backend/requirements.txt
# Core API framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Anthropic Claude — primary AI brain (Phase 3/4/5/6)
anthropic>=0.40.0

# Multi-agent orchestration (Phase 4)
crewai>=0.80.0
crewai-tools>=0.17.0

# Data sources (Phase 2)
yfinance>=0.2.40
pytrends>=4.9.2
praw>=7.7.1
requests>=2.32.0

# Financial data processing (Phase 2+)
pandas>=2.2.0
numpy>=2.0.0

# Vector storage & embeddings (Phase 5/6)
sentence-transformers>=3.0.0
chromadb>=0.5.0

# SEC EDGAR (Phase 2)
httpx==0.27.2
beautifulsoup4==4.12.3
lxml[html_clean]

# Config & utils
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Groq — free-tier fallback for dev/testing
openai==1.57.0

# Flask kept for Phase 4 crew_main standalone runner
flask==3.0.3
flask-cors==4.0.1
